deadlock fix.
shutdown both read and write of the socket to prevent
the reader blocks on read(2) but letting him able to
read pending packets.

foudn when `disconnect` where called and no more packets
are in the socket queue, the reader thread is kept in WAIT
state blocking `stop_reader` function on `pthread_join`.